+++ CHANGELOG.md
@@ -20,6 +20,11 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 - Changed architecture to use one less socket.
+- GUI events are now always handled on a steady timer rather than being
+  interleaved as part of the event loop. This change was made to unify the event
+  handling logic for individually hosted plugins and plugin groups. It should
+  have any negative effects, but please let me know if this does cause unwanted
+  behavior.
 
 ### Fixed